current brightness is always pixelized; initial draft of DecimatingLuminance; add Duration as a type; Serial is non-blocking; wait for Serial enable before proceding

+++ src/Util/ElapsedTime.h
@@ -6,7 +6,7 @@
 
 #include "Misc.h"
 
-typedef long TIME;
+typedef int TIME;
 
 struct ElapsedTime
 {
